Add database/sql like driver API (#18)
* ignore the built binary

* Add database/sql like driver API

This refactors the API for adding drivers . The inspiration is from the `database/sql` . There is a global driver registry inside the `drivers` package, that takes care of storing and retriving drivers.

This brings flexibily for support for more databases since the implementors just need to conform with the interface. This is just the minimun initial API I came up with, we can add more mothods as we seem fit.

*  drivers.Template
I was planning to allow each driver to handle which  templates to generate with the `gnorm init ` command.

I mean something like `gnorm init postgres` will result into sane default templates for postgres that a user will eventualy fine tune if they want to to fit their needs.

I was afraid If I omitted this interface I might forget about it. We can remove it, since I might implement driver specific init templates in a separate PR where this interface is crucial.

* Favor composition over magic

* Move driver interface to database package

* Remove Name  method from the driver interface

This method is not  needed anymore.It is not  up to drivers to assign the names.

* Drop Config.DBType

* Update comment

* Cleanup

- drop unused method
- improve comment
-nchange drivers implementation receivers

* No need for the drivers implementation to be pointers

* Fix typo

* Delete drivers/driver.go

* Delete getDBInfo