Add support for template plugin
This is implementation of the templates functions plugings based on the discussion on the slack channel.

The  design is based on json.

A plugin is a binary/executatble that accepts two aruments , first is the name of the function to invoke, and second is the json payload/context.

The json payload has format `{ "data" : SOME_VALUE }` where SOME_VALUEis any valid json object.

The plugin runner expect the plugin to write to `stdout`  the output after processing the payload, the format of the json output is the same as json input.