enum, json, reformat
THis commit changes the preview function to take an enum rather
than a simple string, since conceptually, an enum is the correct
type, and it gives the caller information about what acceptable
values are.

In addition, add json output.

Finally, reformatted the types output so it exactly mimics the
config settings for type maps, so you can in theory just copy
and paste it, since that's its primary use.