Adds Primary Key Information To Tables and Columns (#50)
* Adds Primary Key Information To Tables and Columns

PrimaryKeys are stored in the data.Table as a slice of columns. IsPrimaryKey is stored in the data.Column as a bool. This allows for easy iteration in template building of all primary key columns of a table.

For Postgres, a new query has been written as queryPrimaryKeys to get the required information. With the required Joins, this could not be done using Gnorm generated types at this time.

For Mysql, the primary key data was already available through the COLUMN_KEY field.

Added Table.HasPrimaryKey()